+++ src/fancy/IMagickImage.h
@@ -30,7 +30,7 @@ public:
   void open(const std::string& filename);
   
   void process(size_t num_colors, 
-                size_t square_side, 
+                unsigned int square_side, 
                 const std::string& suffix);
   
 private:
@@ -47,9 +47,10 @@ private:
   // =================
   // = Inner methods =
   // =================
-  Magick::Color process_element(size_t x0, size_t y0, size_t x1, size_t y1);
+  Magick::Color process_element(unsigned int x0, unsigned int y0,
+                                unsigned int x1, unsigned int y1);
   
-  // static std::vector<Magick::Color> getColors(size_t num_colors, size_t max_color);
+  // static std::vector<Magick::Color> getColors(size_t num_colors, unsigned int max_color);
 };
 
 }}
